requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.25.2
faiss-cpu==1.7.4
langchain==0.1.0
openai==1.3.8
pydantic==2.5.1
python-multipart==0.0.6
pyyaml==6.0.1